requirements.txt
flask
opencv-python
mediapipe
torch
torchvision
numpy
pygame
pillow